refactor: common/utils 폴더 분리
